+++ api/compute/pipeline.go
@@ -27,7 +27,8 @@ import (
 )
 
 // SubmitPipeline executes pipelines using the client and returns the result URI.
-func SubmitPipeline(client *compute.Client, datasets []string, step *pipeline.PipelineDescription, useRunner bool) (string, error) {
+func SubmitPipeline(client *compute.Client, datasets []string,
+	searchRequest *pipeline.SearchSolutionsRequest, step *pipeline.PipelineDescription, useRunner bool) (string, error) {
 
 	config, err := env.LoadConfig()
 	if err != nil {
@@ -45,7 +46,7 @@ func SubmitPipeline(client *compute.Client, datasets []string, step *pipeline.Pi
 
 	request := compute.NewExecPipelineRequest(datasets, step)
 
-	err = request.Dispatch(client)
+	err = request.Dispatch(client, searchRequest)
 	if err != nil {
 		return "", errors.Wrap(err, "unable to dispatch pipeline")
 	}
