+++ api/task/cleaning.go
@@ -72,7 +72,7 @@ func Clean(datasetSource metadata.DatasetSource, schemaFile string, index string
 	// parse primitive response (raw data from the input dataset)
 	// first row of the data is the header
 	// first column of the data is the dataframe index
-	csvData, err := ReadCSVFile(datasetURI, true)
+	csvData, err := util.ReadCSVFile(datasetURI, true)
 	if err != nil {
 		return "", errors.Wrap(err, "unable to parse clean result")
 	}
