+++ api/task/cluster.go
@@ -66,7 +66,7 @@ func ClusterDataset(datasetSource metadata.DatasetSource, schemaFile string, ind
 
 	// open the input file
 	dataPath := path.Join(outputPath.sourceFolder, mainDR.ResPath)
-	lines, err := ReadCSVFile(dataPath, config.HasHeader)
+	lines, err := util.ReadCSVFile(dataPath, config.HasHeader)
 	if err != nil {
 		return "", errors.Wrap(err, "error reading raw data")
 	}
