+++ api/task/pipelines.go
@@ -16,10 +16,7 @@
 package task
 
 import (
-	"encoding/csv"
 	"fmt"
-	"io"
-	"os"
 	"path"
 	"strings"
 
@@ -68,41 +65,6 @@ func SetClient(computeClient *compute.Client) {
 
 func submitPipeline(datasets []string, step *pipeline.PipelineDescription) (string, error) {
 	return sr.SubmitPipeline(client, datasets, step, true)
-}
-
-// ReadCSVFile reads a csv file and returns the string slice representation of the data.
-func ReadCSVFile(filename string, hasHeader bool) ([][]string, error) {
-	// open the file
-	csvFile, err := os.Open(filename)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to open data file")
-	}
-	defer csvFile.Close()
-	reader := csv.NewReader(csvFile)
-
-	lines := make([][]string, 0)
-
-	// skip the header as needed
-	if hasHeader {
-		_, err = reader.Read()
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to read header from file")
-		}
-	}
-
-	// read the raw data
-	for {
-		line, err := reader.Read()
-		if err == io.EOF {
-			break
-		} else if err != nil {
-			return nil, errors.Wrap(err, "failed to read line from file")
-		}
-
-		lines = append(lines, line)
-	}
-
-	return lines, nil
 }
 
 func appendFeature(dataset string, d3mIndexField int, hasHeader bool, feature *FeatureRequest, lines [][]string) ([][]string, error) {
